jf_recomp_resnet50_sun397/requirements.txt
numpy
scipy
argparse
datetime
Pillow
cython
matplotlib
scikit-image
tensorflow==1.12.0
keras==2.1.4
opencv-python
h5py
imgaug
IPython[all]